:wrench: Run test in browser
