+++ .min-wd
@@ -1,0 +1,14 @@
+{
+  "hostname": "localhost",
+  "port": 4444,
+  "sauceLabs": true,
+  "browsers": [
+    {
+      "name": "internet explorer",
+      "version": "10"
+    },
+    {
+      "name": "chrome"
+    }
+  ]
+}
